+++ src/index.ts
@@ -1,3 +1,3 @@
 export { default as Shell } from './Shell';
-export type { ShellTab, ShellProps } from './Shell';
+export type { ShellTab, ShellProps, ShellClassNames } from './Shell';
 export { resolveTab } from './resolveTab';